Add copyright notice to test files
### Explanation

The commit adds a copyright notice to the top of several test files, including `BingProviderTests.cs`, `GoogleProviderTests.cs`, and `UnitTest1.cs`. This change ensures that all test files have a consistent header with copyright information.

### Code Changes

No code changes are necessary for this commit, as the changes are only additions of copyright notices.

```markdown
// Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved. Apache License, Version 2.0
```